% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_noticias_g1.R
\name{baixar_noticias_g1}
\alias{baixar_noticias_g1}
\title{Baixar notícias do g1 a partir das urls obtidas pela função baixar_url_g1}
\usage{
baixar_noticias_g1(urls, diretorio = ".")
}
\arguments{
\item{urls}{Vetor de  urls}

\item{diretorio}{Local onde serão salvas as notícias}
}
\value{
html
}
\description{
Baixar notícias do g1 a partir das urls obtidas pela função baixar_url_g1
}
\examples{
\dontrun{
baixar_noticias_g1(urls = urls)
}
}
