% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_paginas_g1.R
\name{baixar_paginas_g1}
\alias{baixar_paginas_g1}
\title{Baixa páginas do G1}
\usage{
baixar_paginas_g1(numero = 10, diretorio = ".")
}
\arguments{
\item{numero}{Número de páginas a serem baixadas}

\item{diretorio}{Diretório onde salvar as páginas}
}
\value{
html
}
\description{
Baixa páginas do G1
}
\examples{
\dontrun{
baixar_paginas_g1()
}
}
