% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_noticias_g1.R
\name{ler_noticias_g1}
\alias{ler_noticias_g1}
\title{Lê as notícias baixadas pela função baixar_noticias_g1}
\usage{
ler_noticias_g1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos. Se NULL, informar diretório}

\item{diretorio}{Informar diretório onde estão os arquivos}
}
\value{
tibbe
}
\description{
Lê as notícias baixadas pela função baixar_noticias_g1
}
\examples{
\dontrun{
df <- ler_noticias_g1()
}
}
