% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_urls_g1.R
\name{ler_urls_g1}
\alias{ler_urls_g1}
\title{Lê as urls contidas nas páginas baixadas por baixar_paginas_g1}
\usage{
ler_urls_g1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se NULL, informar diretório}

\item{diretorio}{Indicar diretório se não indicou arquivos}
}
\value{
character vector
}
\description{
Lê as urls contidas nas páginas baixadas por baixar_paginas_g1
}
\examples{
\dontrun{
urls <- ler_urls_g1(diretorio=".")
}
}
